requirements.txt
streamlit
tensorflow-cpu
librosa
numpy>=2.1.0
pandas
scikit-learn
matplotlib
seaborn
spotipy
yt-dlp
requests
soundfile
python-dotenv